Commit (broken) first attempt; changing direction
Commit current work on implementation of tic-tac-toe algorithm that I
got from a description on some random website. In my tests, it seems to
be failing and allowing the user to win. In retrospect this was a
bad idea; I should not have trusted it to work. Then again maybe I have
not understood the restrictions on that site.

I could go through and do debugging to add the cases that cause it to
fail, but the more I work on this the more I'd feel comfortable doing a
general MiniMax implementation instead, and this just cements it. It
would better show off skills to the reviewer, and it would be easier to
adapt to the task if we pretend that this was something more complex
than implementing tic-tac-toe, and I could be fairly certain that the
solution would be optimal.