Implement naive minimax (negamax) alg.; working.
With this commit, the negamax algorithm is coded and working, with some
tests implemented. The recursive step is quite unnecessarily heavy; I
would implement some of my planned enhancements before calling it
finished.

Next step should probably be to hook this module up to a django view to
expose the game to the web.